+++ src/JK.Common/CHANGELOG.md
@@ -13,6 +13,14 @@ Unless otherwise noted, all changes made by [@jeremyknight-me](https://github.co
 
 - Added `Parse` methods to `SpanExtensions`
 - Added `Parse` methods to `StringExtensions`
+- Added `Range` and `Index` "polyfills"
+- Added Regex code generation to `RegexHelper`
+- Added `HasItem` extension methods for `ICollection<T>` and `IReadOnlyCollection<T>`
+
+### Changed
+
+- Changed `DateTimeOffsetFactory` to static class.
+- Changed `StringHelper.Right` to use range operator
 
 ## 4.0.0